bullbear/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
langgraph>=0.0.20
langchain>=0.1.0
pathway>=0.7.0
pandas>=2.0.0
requests>=2.31.0
